refactor: 추천 캐러셀 컴포넌트 코드 분리
